move shared secret to QRCodeData
